Move test to wren-test
