+++ wren-tests/pom.xml
@@ -125,6 +125,11 @@
         </dependency>
 
         <dependency>
+            <groupId>io.wren</groupId>
+            <artifactId>wren-sqlglot-converter</artifactId>
+        </dependency>
+
+        <dependency>
             <groupId>javax.ws.rs</groupId>
             <artifactId>javax.ws.rs-api</artifactId>
         </dependency>
